refactor: move conflicts on locale_based_test.dart
